requirements_medgemma.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.2+cu118
torchvision==0.16.2+cu118
torchaudio==2.1.2+cu118

# Transformers ecosystem — MedGemma requires >=4.47, pin to tested version
transformers==4.50.0
tokenizers==0.21.0
accelerate==0.34.2
peft==0.13.2
trl==0.12.1
bitsandbytes==0.44.1
safetensors==0.4.5
huggingface-hub==0.26.5

# Distributed training
deepspeed==0.14.2

# Data
datasets==2.21.0
pandas>=2.0.0
numpy>=1.24.0,<2.0

# Medical entity extraction
radgraph

# OpenCLIP (installed editable from ./train/open_clip)
timm==1.0.12
einops>=0.7.0
ftfy>=6.0.0
braceexpand
webdataset>=0.2.0

# Evaluation
evaluate>=0.4.0
rouge-score>=0.1.2
sacrebleu>=2.3.0
scikit-learn>=1.2.0
scipy>=1.10.0

# Vision
Pillow>=10.0.0

# Utilities — unpinned so PyPI resolver wins over the torch index
tqdm>=4.65.0
wandb>=0.16.0
PyYAML>=6.0
sentencepiece>=0.1.99
protobuf>=4.21.0

# Kaggle / notebook
pyngrok>=7.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=8.0.0
